♻️ made changes based on review comments
